% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsarif-package.R
\docType{package}
\name{testsarif-package}
\alias{testsarif}
\alias{testsarif-package}
\title{testsarif: A Test for 'jarl'}
\description{
Just a test.
}
\author{
\strong{Maintainer}: John Doe \email{john.doe@example.com}

Authors:
\itemize{
  \item John Doe \email{john.doe@example.com}
}

}
\keyword{internal}
